implement modules to store and retrieve from localstorage
