+++ src/getlist.js
@@ -1,0 +1,2 @@
+const getList = localStorage.getItem('todo') ? JSON.parse(localStorage.getItem('todo')) : [];
+export default getList;
