+++ src/storelist.js
@@ -1,0 +1,5 @@
+const storeList = (l) => {
+  localStorage.setItem('todo',
+    JSON.stringify(l.sort((a, b) => a.index - b.index)));
+};
+export default storeList;
